2.1a1 - Commit final

+++ api/renderer/homepage.php
@@ -1,36 +1,31 @@
 <?php
 
 function getAvgLuminance($filename, $num_samples=30) {
-    // needs a mimetype check
-    $img = imagecreatefromjpeg($filename);
-    $width = imagesx($img);
-    $height = imagesy($img);
-    $x_step = intval($width/$num_samples);
-    $y_step = intval($height/$num_samples);
-    $total_lum = 0;
-    $sample_no = 1;
-    for ($x=0; $x<$width; $x+=$x_step) {
-        for ($y=0; $y<$height; $y+=$y_step) {
-            $rgb = imagecolorat($img, $x, $y);
-            $r = ($rgb >> 16) & 0xFF;
-            $g = ($rgb >> 8) & 0xFF;
-            $b = $rgb & 0xFF;
-            // choose a simple luminance formula from here
-            // http://stackoverflow.com/questions/596216/formula-to-determine-brightness-of-rgb-color
-            $lum = ($r+$r+$b+$g+$g+$g)/6;
-            $total_lum += $lum;
-            $sample_no++;
+    if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/cache/banner.mtd")) {
+        return file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/cache/banner.mtd");
+    } else {
+        $img = imagecreatefromjpeg($filename);
+        $width = imagesx($img);
+        $height = imagesy($img);
+        $x_step = intval($width/$num_samples);
+        $y_step = intval($height/$num_samples);
+        $total_lum = 0;
+        $sample_no = 1;
+        for ($x=0; $x<$width; $x+=$x_step) {
+            for ($y=0; $y<$height; $y+=$y_step) {
+                $rgb = imagecolorat($img, $x, $y);
+                $r = ($rgb >> 16) & 0xFF;
+                $g = ($rgb >> 8) & 0xFF;
+                $b = $rgb & 0xFF;
+                $lum = ($r+$r+$b+$g+$g+$g)/6;
+                $total_lum += $lum;
+                $sample_no++;
+            }
         }
+        $avg_lum  = $total_lum / $sample_no;
+        file_put_contents($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/cache/banner.mtd", ($avg_lum / 255) * 100);
+        return ($avg_lum / 255) * 100;
     }
-    // work out the average
-    $avg_lum  = $total_lum / $sample_no;
-    return ($avg_lum / 255) * 100;
-}
-
-function compareASCII($a, $b) {
-    $at = iconv('UTF-8', 'ASCII//TRANSLIT', $a);
-    $bt = iconv('UTF-8', 'ASCII//TRANSLIT', $b);
-    return strcmp($at, $bt);
 }
 
 ?>
@@ -42,8 +37,6 @@ if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent")) {
 } else {
     $ready = false;
 }
-
-// $ready = false; // Switch to 'false' to test the OOBE
 
 ?>
 
@@ -68,33 +61,40 @@ if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent")) {
     <link rel="stylesheet" href="/resources/css/fonts-import.css">
     <link rel="stylesheet" href="/resources/css/ui.css">
     <title><?php
-    
+
+    function compareASCII($a, $b) {
+        $at = iconv('UTF-8', 'ASCII//TRANSLIT', $a);
+        $bt = iconv('UTF-8', 'ASCII//TRANSLIT', $b);
+        return strcmp($at, $bt);
+    }
+
     if ($ready) {
         echo(file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/sitename"));
     } else {
-        echo("Votre site est presque prêt - MPCMS");
+        echo("MPCMS");
     }
 
     ?></title>
     <?php
-    
+
     if ($ready) {
         include_once $_SERVER['DOCUMENT_ROOT'] . "/resources/private/header.php";
     }
-    
+
     ?>
 </head>
 <body>
     <?php
 
     if (!$ready) {
-        die('<div class="centered discover"><h1>Minteck Projects CMS</h1><h3>La nouvelle génération de sites Web</h3><p>Tous les fichiers ont été copiés correctement, vous devez maintenant configurer le logiciel Minteck Projects CMS.</p><p>Pour cela, nous vous conseillons d\'utiliser un ordinateur, ou tout autre appareil avec un écran plus grand.</p><a href="/cms-special/setup"><img src="/resources/image/config_explore.svg">Configurer</a><br><br><hr><small>' . file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/api/version") . '</small></div>');
+        // die('<div class="centered discover"><h1>Minteck Projects CMS</h1><h3>La nouvelle génération de sites Web</h3><p>Tous les fichiers ont été copiés correctement, vous devez maintenant configurer le logiciel Minteck Projects CMS.</p><p>Pour cela, nous vous conseillons d\'utiliser un ordinateur, ou tout autre appareil avec un écran plus grand.</p><a href="/cms-special/setup"><img src="/resources/image/config_explore.svg">Configurer</a><br><br><hr><small>' . file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/api/version") . '</small></div>');
+        die("<script>location.href='/cms-special/setup';</script>");
     }
-    
+
     if ($ready) {
         echo("<script type=\"text/javascript\">\nvar pushbar = new Pushbar({\nblur:true,\noverlay:true,\n});\n</script>");
     }
-    
+
 
     if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/resources/upload/banner.jpg")) {
         $banner = "/resources/upload/banner.jpg";
@@ -121,7 +121,7 @@ if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent")) {
     </div>
     <div id="menubar"><span class="menubar-link menubar-mobile" id="menubar-link-navigation" onclick="pushbar.open('panel-navigation')"><img src="/resources/image/menu.svg" class="menubar-img"><span class="menubar-link-text"><?= $lang["viewer"]["menu"] ?></span></span>
         <?php
-        
+
         if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/alwaysmenu")) {
             echo('<span class="menubar-link menubar-desktop" id="menubar-link-navigation" onclick="pushbar.open(\'panel-navigation\')"><img src="/resources/image/menu.svg" class="menubar-img"><span class="menubar-link-text">' . $lang["viewer"]["menu"] . '</span></span>');
         } else {
@@ -129,28 +129,32 @@ if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent")) {
             echo('<a href="/" title="/" class="menulink-desktop">' . $lang["viewer"]["home"] . '</a>');
             $count = $count + 1;
 
-            $pages = scandir($_SERVER['DOCUMENT_ROOT']);
-            uasort($pages, 'compareASCII');
-            foreach ($pages as $page) {
-                if ($page != ".." && $page != ".") {
-                    if (is_dir($_SERVER['DOCUMENT_ROOT'] . "/" . $page)) {
-                        if ($count < 4) {
-                            if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/" . $page . "/pagename")) {
-                                if (!in_array($page, $customSettings->PagesMasquées)) {
-                                    echo("<a href=\"/{$page}\" title=\"/{$page}\" class=\"menulink-desktop\">" . file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/" . $page . "/pagename") . "</a>");
-                                    $count = $count + 1;
+            if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/cache/pagelist.mtd")) {
+                echo(file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/cache/pagelist.mtd"));
+            } else {
+                $pages = scandir($_SERVER['DOCUMENT_ROOT']);
+                uasort($pages, 'compareASCII');
+                foreach ($pages as $page) {
+                    if ($page != ".." && $page != ".") {
+                        if (is_dir($_SERVER['DOCUMENT_ROOT'] . "/" . $page)) {
+                            if ($count < 4) {
+                                if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/" . $page . "/pagename")) {
+                                    if (!in_array($page, $customSettings->PagesMasquées)) {
+                                        echo("<a href=\"/{$page}\" title=\"/{$page}\" class=\"menulink-desktop\">" . file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/" . $page . "/pagename") . "</a>");
+                                        $count = $count + 1;
+                                    }
                                 }
                             }
                         }
                     }
                 }
+                if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/galery/enabled")) {echo("<a href=\"/cms-special/galery\" title=\"/cms-special/galery\" class=\"menulink-desktop\">" . $lang["viewer"]["galery"] . "</a>");$count = $count + 1;}
             }
-            if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/galery/enabled")) {echo("<a href=\"/cms-special/galery\" title=\"/cms-special/galery\" class=\"menulink-desktop\">" . $lang["viewer"]["galery"] . "</a>");$count = $count + 1;}
             if ($count >= 4) {
                 echo("<a onclick=\"pushbar.open('panel-navigation')\" title=\"" . $lang["viewer"]["menutitle"] . "\" class=\"menulink-desktop\">" . $lang["viewer"]["menudesktop"] . "</a>");
             }
         }
-        
+
         ?>
         <?php
 
@@ -163,7 +167,7 @@ if (!empty($widgets->list)) {
     <div data-pushbar-id="panel-navigation" class="pushbar from_left">
         <div id="banner-menu" style='background-image: url("<?= $banner ?>");'>
             <img id="banner-menu-logo" src="/resources/upload/siteicon.png"><span id="banner-menu-name" <?php if ($blackBannerText) {echo("class=\"banner-black\"");} ?>><?php
-            
+
             $sitename = file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/sitename");
 
             if (strlen($sitename) < 15) {
@@ -171,7 +175,7 @@ if (!empty($widgets->list)) {
             } else {
                 echo(substr($sitename, 0, 14) . "...");
             }
-            
+
             ?></span>
         </div>
         <img src="/resources/image/close.svg" id="menubar-close" onclick="pushbar.close()">
@@ -223,7 +227,7 @@ if (!empty($widgets->list)) {
     <div id="page-placeholder">
         <div id="page-content">
             <?php
-            
+
             $html_string = file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/pages/index");
             preg_match_all('#<h[1-6]*[^>]*>.*?<\/h[1-6]>#',$html_string,$results);
 
@@ -238,18 +242,18 @@ if (!empty($widgets->list)) {
             $toc = preg_replace('#<\/h5>#','</li>',$toc);
             $toc = preg_replace('#<h6>#','<li class="toc$1" style="margin-left: 80px;">',$toc);
             $toc = preg_replace('#<\/h6>#','</li>',$toc);
-        
-            $toc = '<div id="toc"> 
+
+            $toc = '<div id="toc">
             <h3>' . $lang["viewer"]["toc"] . '</h3>
             <ul>
             '.$toc.'
             </ul>
             </div><hr>';
-        
+
             if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/semantic_toc")) {
                 echo($toc);
             }
-            
+
             ?>
             <?php echo($html_string); ?>
         </div>
